% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timetable.R
\name{plot_timetable}
\alias{plot_timetable}
\title{Plot a weekly timetable (created by Peng Zhao)}
\usage{
plot_timetable(
  timetable,
  caption = "Peng Zhao's timetable drawn with the R langauge."
)
}
\arguments{
\item{timetable}{a data frame or the path for a .csv file}

\item{caption}{title of the timetable}
}
\value{
a ggplot image
}
\description{
Plot a weekly timetable (created by Peng Zhao)
}
\examples{
library(ggplot2)
dtf <- structure(list(weekday = c(2L, 2L), hour_from = 10:11, hour_to = c(10.25,
    12), event = c("Coffee break", "Staff Seminar"), location = c("ES316", "ES354"),
    color = c("brown", "red")), row.names = 3:4, class = "data.frame")
plot_timetable(dtf)
ggsave("image/weekly.pdf", device = cairo_pdf, width = 297, height = 210, units = "mm")

plot_timetable(timetable = "overview/timetable.csv") + theme_void()
}
